+++ Frends.AzureBlobStorage.UploadBlob/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [3.0.0] - 2025-04-16
+### Changed
+- [Breaking] Reorganized and renamed parameters for clarity and consistency
+  - To upgrade to the new version you can select the new parameters matching the old ones.  
+    You can find a list of the changes to parameter locations and names below:
+    - Renamed Source parameter tab to Input
+    - Renamed Destination parameter tab to Connection
+    - Destination.HandleExistingFile renamed and moved to Input.ActionOnExistingFile
+    - HandleExistingFile.Error renamed to OnExistingFile.Throw
+    - ConnectionMethod.SASToken renamed to ConnectionMethod.SasToken
+    - Destination.ApplicationID renamed and moved to Connection.ApplicationId
+    - Destination.TenantID renamed and moved to Connection.TenantId
+    - Destination.SASToken renamed and moved to Connection.SasToken
+    - Destination.BlobType moved to Options.BlobType
+    - Destination.ResizeFile moved to Options.ResizeFile
+    - Destination.PageMaxSize moved to Options.PageMaxSize
+    - Destination.PageOffset moved to Options.PageOffset
+    - Destination.ContentType moved to Options.ContentType
+    - Destination.Encoding moved to Options.Encoding
+    - FileEncoding.WINDOWS1252 renamed to Windows1252
+    - Destination.EnableBOM renamed and moved to Options.EnableBom
+    - Destination.FileEncodingString moved to Options.FileEncodingString
+    - Destination.ParallelOperations moved to Options.ParallelOperations
+
 ## [2.4.0] - 2025-01-27
 ### Added
 - Added SAS Token authentication method.
@@ -22,10 +46,10 @@
 - Azure.Storage.Blobs to version 12.19.1
 
 ## [2.0.0] - 2023-04-06
-##Added:
+### Added
 Option to choose whether to upload a directory or just a single blob.
 New parameters: Destination.ResizeFile, Source.SourceType, Source.SourceDirectory, Source.SearchPattern, Source.BlobName, Source.BlobFolderName.
-##Changed:
+### Changed
 Destination.BlobName and Destination.RenameTo parameters have been replaced by Source.BlobName or Source.BlobFolderName depending on Source.SourceType.
 Result.SourceFile and Result.Uri have been replaced by Result.Data.